[WIP] all nodes are now classes
